Codes/ci/requirements.txt
robotframework==3.0.2
robotframework-selenium2library==1.8.0
paramiko==2.3.1
pf
nfs
requests
tornado
docopt
pyyaml
ddc